pyproject.toml
[tool.poetry]
name = "hr-workers"
version = "0.1.0"
description = ""
authors = ["IvanovSemjon <IvanovSemjon@yabdex.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
django = "^4.2"
djangorestframework = "^3.16.1"
openpyxl = "^3.1.5"
psycopg2-binary = "^2.9.11"
django-filter = "<25.0"
python-dotenv = "^1.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-django = "^4.11.1"
black = "^25.9.0"
isort = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

